Telemetry: Known logger reporting
This introduces telemetry around the `loggerName` when you call
`shouldLog`. This telemetry only sends the `loggerName` and the number
of times it was called at each level. This is reported periodically
over time.

The telemetry data is used to populate the logger hierarchy in the
Prefab UI.

You can disable the telemetry by setting `collectLoggerCounts: false`
when initializing.